updated app.js
