--- client/src/components/Home.js
@@ -1,11 +1,0 @@
-import React, { Component } from 'react'
-
-export default class Home extends Component {
-    render() {
-        return (
-            <div>
-                <p>This is home page</p>
-            </div>
-        )
-    }
-}
