+++ client/src/pages/Home.js
@@ -1,80 +1,15 @@
-import React, { useState } from "react";
+import React, { Component } from 'react'
 import "../App.css";
-import { Container, Row, Button, Col } from "react-bootstrap";
 import Polly from '../components/Polly'
 import Axios from "axios";
 
-function Home(props) {
 
- 
-  return (
+export default class Home extends Component {
+  render() {
+    return (
+      <div>
 
-    <Container className="signup">
-      <Row>
-        <Col md={{ span: 8, offset: 2 }}>
-          <h1>Home Page</h1>
-          {isAuth ? (
-            <>
-              <Button
-                className="m-1"
-                onClick={e => {
-                  e.preventDefault();
-                  setSecret('');
-                  logout();
-                }}
-              >
-                Logout
-              </Button>
-              <Button
-                className="m-1"
-                onClick={e => {
-                  e.preventDefault();
-                  props.history.push("/members");
-                }}
-              >
-                Members
-              </Button>
-            </>
-          ) : (
-              <>
-                <Button
-                  className="m-1"
-                  onClick={e => {
-                    e.preventDefault();
-                    props.history.push("/login");
-                  }}
-                >
-                  Login
-              </Button>
-                <Button
-                  className="m-1"
-                  onClick={e => {
-                    e.preventDefault();
-                    props.history.push("/signup");
-                  }}
-                >
-                  Signup
-              </Button>
-              </>
-            )}
-          <Button
-            className="m-1"
-            onClick={e => {
-              e.preventDefault();
-              getSecret();
-            }}
-          >
-            Show Secrets
-          </Button>
-        </Col>
-      </Row>
-      <Row>
-        <Col md={{ span: 8, offset: 2 }}>
-          <h1>{secret}</h1>
-        </Col>
-      </Row>
-    </Container>
-  );
+      </div>
+    )
+  }
 }
-
-export default Home;
